消除一个警告

+++ MyWeiboJingXuan/MyWeiboJingXuan/Classes/Category/NSString+XFTime.m
@@ -9,11 +9,13 @@
 #import "NSString+XFTime.h"
 
 @implementation NSString (XFTime)
+
 + (NSString *)stringWithTime:(NSTimeInterval)time {
     
-    NSInteger min = time / 60;
+    NSInteger min = (NSInteger)time / 60;
     NSInteger second = (NSInteger)time % 60;
     
-    return [NSString stringWithFormat:@"%02ld:%02ld", min, second];
+    return [NSString stringWithFormat:@"%02ld:%02ld", (long)min, (long)second];
 }
+
 @end
